Refactor code, improve javadoc and UG and PPP
