+++ docs/team/tamkenghong.md
@@ -5,14 +5,15 @@ title: Tam Keng Hong's Project Portfolio Page
 
 ### Overview
 
-To be added soon.
+Draft Draft.
 
 ### Summary of Contributions
 
-- Code contributed
-- Enhancements implemented: A summary of the enhancements you implemented.
-- Contributions to the UG: Which sections did you contribute to the UG?
-- Contributions to the DG: Which sections did you contribute to the DG? Which UML diagrams did you add/updated?
-- Contributions to team-based tasks
-- Review/mentoring contributions: Links to PRs reviewed, instances of helping team members in other ways.
-- Contributions beyond the project team:
+- Code contributed: Created classes and methods for: Assignment, AssignmentList, Submissions, and 
+everything related to them (Parser, command, testcases and so on.) 
+- Enhancements implemented: For assignments: add_asgn, delete_asgn, grade, ungrade, list_asgn.
+- Contributions to the UG: Everything related to assignments and submissions.
+- Contributions to the DG: Everything related to assignments and submissions.
+- Contributions to team-based tasks: (TODO: Draft)
+- Review/mentoring contributions: Links to PRs reviewed, instances of helping team members in other ways. (TODO: Draft)
+- Contributions beyond the project team: Helped with design (Logo + first draft screenshot).
