+++ src/main/java/taa/model/ClassList.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import java.util.List;
 
 import javafx.collections.ObservableList;
-import taa.assignment.AssignmentList;
+import taa.model.assignment.AssignmentList;
 import taa.model.student.Student;
 import taa.model.student.UniqueStudentList;
