+++ src/main/java/taa/model/ModelManager.java
@@ -38,12 +38,6 @@ import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import javafx.scene.control.Alert;
 import javafx.util.Duration;
-import taa.assignment.AssignmentList;
-import taa.assignment.exceptions.AssignmentException;
-import taa.assignment.exceptions.AssignmentNotFoundException;
-import taa.assignment.exceptions.DuplicateAssignmentException;
-import taa.assignment.exceptions.NoGradeVarianceException;
-import taa.assignment.exceptions.NoSubmissionsFoundException;
 import taa.commons.core.GuiSettings;
 import taa.commons.core.LogsCenter;
 import taa.commons.core.index.Index;
@@ -51,6 +45,12 @@ import taa.commons.util.CollectionUtil;
 import taa.logic.commands.enums.ChartType;
 import taa.model.alarm.Alarm;
 import taa.model.alarm.AlarmList;
+import taa.model.assignment.AssignmentList;
+import taa.model.assignment.exceptions.AssignmentException;
+import taa.model.assignment.exceptions.AssignmentNotFoundException;
+import taa.model.assignment.exceptions.DuplicateAssignmentException;
+import taa.model.assignment.exceptions.NoGradeVarianceException;
+import taa.model.assignment.exceptions.NoSubmissionsFoundException;
 import taa.model.student.Attendance;
 import taa.model.student.Name;
 import taa.model.student.SameStudentPredicate;
