+++ src/main/java/taa/model/assignment/AssignmentList.java
@@ -1,25 +1,29 @@
-package taa.assignment;
+package taa.model.assignment;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 
 import javafx.collections.transformation.FilteredList;
-import taa.assignment.exceptions.AssignmentNotFoundException;
-import taa.assignment.exceptions.DuplicateAssignmentException;
-import taa.assignment.exceptions.InvalidGradeException;
-import taa.assignment.exceptions.SubmissionNotFoundException;
+import taa.model.assignment.exceptions.AssignmentNotFoundException;
+import taa.model.assignment.exceptions.DuplicateAssignmentException;
+import taa.model.assignment.exceptions.InvalidGradeException;
+import taa.model.assignment.exceptions.SubmissionNotFoundException;
 import taa.model.student.Student;
 
 /**
- * List of assignments
+ * List of assignments.
+ * This is the class that the model uses to do all things related to assignments and submissions.
  */
 public class AssignmentList {
     private final ArrayList<Assignment> assignments = new ArrayList<>();
     private final HashMap<String, Assignment> assignmentMap = new HashMap<>();
 
     /**
+     * Adds an assignment to the AssignmentList.
      * @param assignmentName
      * @param sl
+     * @param totalMarks
+     * @throws DuplicateAssignmentException if an assignment with assignmentName already exists
      */
     public void add(String assignmentName, FilteredList<Student> sl, int totalMarks)
             throws DuplicateAssignmentException {
@@ -33,7 +37,9 @@ public class AssignmentList {
     }
 
     /**
+     * Deletes an assignment from the AssignmentList.
      * @param assignmentName
+     * @throws AssignmentNotFoundException if an assignment with assignmentName does not exist in the AssignmentList
      */
     public void delete(String assignmentName) throws AssignmentNotFoundException {
         if (!assignmentMap.containsKey(assignmentName)) {
@@ -46,10 +52,15 @@ public class AssignmentList {
     }
 
     /**
+     * Grades a student submission for an assignment of assignmentName.
+     * The submission will be updated according to the marks given and whether it is a late submission.
      * @param assignmentName
      * @param student
      * @param marks
      * @param isLateSubmission
+     * @throws AssignmentNotFoundException
+     * @throws SubmissionNotFoundException
+     * @throws InvalidGradeException
      */
     public void grade(String assignmentName, Student student, int marks, boolean isLateSubmission)
             throws AssignmentNotFoundException, SubmissionNotFoundException, InvalidGradeException {
@@ -61,6 +72,7 @@ public class AssignmentList {
     }
 
     /**
+     * Resets the marks and late status of a student submission for an assignment of assignmentName
      * @param assignmentName
      * @param student
      * @throws AssignmentNotFoundException
@@ -76,7 +88,8 @@ public class AssignmentList {
     }
 
     /**
-     * @return A list of all assignments and submissions
+     * Lists out all the assignments and student submissions to the command line interface.
+     * @return A string, which is the list of all assignments and submissions.
      */
     public String list() {
         StringBuilder sb = new StringBuilder();
@@ -90,7 +103,8 @@ public class AssignmentList {
     }
 
     /**
-     * Delete all of a student's submissions from the assignment list.
+     * When a student is deleted, all the student's submissions must be deleted as well.
+     * This deletes all of a student's submissions from the AssignmentList.
      * @param s the student
      */
     public void deleteStudent(Student s) {
@@ -100,7 +114,8 @@ public class AssignmentList {
     }
 
     /**
-     * Creates new submissions for an added student.
+     * When a student is added, all the existing assignments should be linked to the new student as well.
+     * This creates new submissions for an added student for all the existing assignments.
      * @param s the student
      */
     public void addStudent(Student s) {
@@ -110,7 +125,9 @@ public class AssignmentList {
     }
 
     /**
-     * Populates the assignment list, submissions list into
+     * On startup, this will populate the assignment list and submissions from the
+     * submission storage string data held by each student.
+     * This is also called when we edit a student.
      * @param sl the student list
      */
     public void initFromStorage(FilteredList<Student> sl) {
@@ -141,7 +158,8 @@ public class AssignmentList {
     }
 
     /**
-     * Returns true if an assignment with the provided name exists.
+     * @param name
+     * @return true if an assignment with the provided name exists.
      */
     public boolean contains(String name) {
         return this.assignmentMap.containsKey(name);
