+++ src/main/java/taa/ui/StudentCard.java
@@ -7,7 +7,7 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import taa.assignment.Submission;
+import taa.model.assignment.Submission;
 import taa.model.student.Student;
 
 /**
